Migrate action to TypeScript and restore preview coverage
