+++ .github/workflows/test.yml
@@ -4,27 +4,55 @@ on:
   workflow_dispatch:
   pull_request:
   push:
-    branches: master
-
-permissions:
-  actions: write
-  contents: read
+    branches: main
 
 jobs:
-  test:
+  quality:
     runs-on: ubuntu-latest
-    name: HTML Preview
+    name: Quality Checks
+    permissions:
+      contents: read
     steps:
-      - uses: actions/checkout@v4
-      - uses: oven-sh/setup-bun@v2
+      - uses: actions/checkout@v6
+      - uses: pnpm/action-setup@v4
         with:
-          bun-version: 1.3.10
+          version: 10
+          cache: true
+      - uses: actions/setup-node@v6
+        with:
+          node-version: 24
+          cache: pnpm
       - name: Install dependencies
-        run: bun install --frozen-lockfile
+        run: pnpm install --frozen-lockfile
       - name: Run unit tests
-        run: bun run test
+        run: pnpm test
+      - name: Run typecheck
+        run: pnpm typecheck
+      - name: Run lint
+        run: pnpm lint
       - name: Build action bundle
-        run: bun run build
+        run: pnpm build
+
+  preview_single_file:
+    runs-on: ubuntu-latest
+    name: Preview Single File
+    permissions:
+      actions: write
+      contents: read
+    steps:
+      - uses: actions/checkout@v6
+      - uses: pnpm/action-setup@v4
+        with:
+          version: 10
+          cache: true
+      - uses: actions/setup-node@v6
+        with:
+          node-version: 24
+          cache: pnpm
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+      - name: Build action bundle
+        run: pnpm build
       - name: Prepare generated single-file site
         run: |
           mkdir -p .tmp/generated-single
@@ -43,6 +71,27 @@ jobs:
           test "${{ steps.single_file_preview.outputs.artifact_count }}" = "1"
           test "${{ steps.single_file_preview.outputs.discovered_files_count }}" = "1"
           test -n "${{ steps.single_file_preview.outputs.artifact_url }}"
+
+  preview_multi_file_default:
+    runs-on: ubuntu-latest
+    name: Preview Multi File Default
+    permissions:
+      actions: write
+      contents: read
+    steps:
+      - uses: actions/checkout@v6
+      - uses: pnpm/action-setup@v4
+        with:
+          version: 10
+          cache: true
+      - uses: actions/setup-node@v6
+        with:
+          node-version: 24
+          cache: pnpm
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+      - name: Build action bundle
+        run: pnpm build
       - name: Prepare generated multi-file site
         run: |
           mkdir -p .tmp/generated-site
@@ -57,11 +106,36 @@ jobs:
       - name: Assert multi-file preview outputs
         run: |
           test "${{ steps.generated_site_preview.outputs.mode }}" = "artifact"
-          test "${{ steps.generated_site_preview.outputs.upload_strategy }}" = "multi_raw"
-          test "${{ steps.generated_site_preview.outputs.artifact_count }}" = "6"
+          test "${{ steps.generated_site_preview.outputs.upload_strategy }}" = "archive"
+          test "${{ steps.generated_site_preview.outputs.artifact_count }}" = "1"
           test "${{ steps.generated_site_preview.outputs.discovered_files_count }}" = "6"
           test -n "${{ steps.generated_site_preview.outputs.artifact_url }}"
           test -n "${{ steps.generated_site_preview.outputs.artifact_manifest }}"
+
+  preview_multi_file_archive:
+    runs-on: ubuntu-latest
+    name: Preview Multi File Archive
+    permissions:
+      actions: write
+      contents: read
+    steps:
+      - uses: actions/checkout@v6
+      - uses: pnpm/action-setup@v4
+        with:
+          version: 10
+          cache: true
+      - uses: actions/setup-node@v6
+        with:
+          node-version: 24
+          cache: pnpm
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+      - name: Build action bundle
+        run: pnpm build
+      - name: Prepare generated multi-file site
+        run: |
+          mkdir -p .tmp/generated-site
+          cp -R test/fixtures/linked-pages/. .tmp/generated-site/
       - name: Upload generated multi-file preview as archive
         id: generated_site_archive_preview
         uses: ./
@@ -77,6 +151,26 @@ jobs:
           test "${{ steps.generated_site_archive_preview.outputs.artifact_count }}" = "1"
           test "${{ steps.generated_site_archive_preview.outputs.discovered_files_count }}" = "6"
           test -n "${{ steps.generated_site_archive_preview.outputs.artifact_url }}"
+
+  preview_repo_mode:
+    runs-on: ubuntu-latest
+    name: Preview Repo Mode
+    permissions:
+      contents: read
+    steps:
+      - uses: actions/checkout@v6
+      - uses: pnpm/action-setup@v4
+        with:
+          version: 10
+          cache: true
+      - uses: actions/setup-node@v6
+        with:
+          node-version: 24
+          cache: pnpm
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+      - name: Build action bundle
+        run: pnpm build
       - name: Validate repo mode still works
         id: repo_preview
         uses: ./
