+++ package.json
@@ -4,17 +4,30 @@
   "description": "GitHub action to upload preview artifacts for HTML files and generated sites",
   "private": true,
   "type": "module",
-  "main": "dist/index.js",
-  "packageManager": "bun@1.3.10",
+  "main": "dist/index.mjs",
+  "packageManager": "pnpm@10.30.3",
   "scripts": {
-    "build": "bun build ./index.js --target=node --outfile ./dist/index.js",
-    "package": "bun run build",
-    "test": "node --test"
+    "build": "tsdown",
+    "package": "pnpm run build",
+    "test": "tsx --test test/**/*.test.ts",
+    "typecheck": "tsc --noEmit",
+    "lint": "oxlint src test",
+    "lint:fix": "oxlint --fix src test",
+    "format": "oxfmt src/**/*.ts test/**/*.ts tsdown.config.ts",
+    "format:check": "oxfmt --check src/**/*.ts test/**/*.ts tsdown.config.ts"
   },
   "license": "MIT",
   "dependencies": {
     "@actions/artifact": "^6.2.0",
     "@actions/core": "^3.0.0",
     "@actions/github": "^9.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^24.0.0",
+    "oxfmt": "^0.36.0",
+    "oxlint": "^1.51.0",
+    "tsdown": "^0.20.3",
+    "tsx": "^4.21.0",
+    "typescript": "^5.9.3"
   }
 }
