+++ test/fixtures/basic-single-file/index.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>Basic Preview</title>
   </head>
   <body>
