+++ test/fixtures/discovery-escape-attempt/index.html
@@ -1,15 +1,15 @@
 <!doctype html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>Escape Attempt</title>
-    <link rel="stylesheet" href="https://example.com/app.css">
+    <link rel="stylesheet" href="https://example.com/app.css" />
   </head>
   <body>
     <a href="../secret.txt">Outside Root</a>
     <a href="nested/inside.html#section">Nested</a>
     <a href="mailto:test@example.com">Mail</a>
     <a href="javascript:void(0)">JS</a>
-    <img src="/images/safe.png" alt="Safe">
+    <img src="/images/safe.png" alt="Safe" />
   </body>
 </html>
