+++ test/fixtures/discovery-escape-attempt/nested/inside.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>Inside</title>
   </head>
   <body>
