+++ test/fixtures/linked-pages/about.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>About</title>
   </head>
   <body>
