+++ test/fixtures/linked-pages/assets/site.css
@@ -1,2 +1,6 @@
-body { font-family: sans-serif; }
-.hero { background-image: url('../images/logo.svg'); }
+body {
+  font-family: sans-serif;
+}
+.hero {
+  background-image: url("../images/logo.svg");
+}
