+++ test/fixtures/linked-pages/assets/team.css
@@ -1,1 +1,3 @@
-body { color: #333; }
+body {
+  color: #333;
+}
