+++ test/fixtures/linked-pages/index.html
@@ -1,12 +1,12 @@
 <!doctype html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>Linked Pages</title>
-    <link rel="stylesheet" href="/assets/site.css">
+    <link rel="stylesheet" href="/assets/site.css" />
   </head>
   <body>
-    <img src="images/logo.svg" alt="Logo">
+    <img src="images/logo.svg" alt="Logo" />
     <a href="about.html">About</a>
     <a href="nested/team.html">Team</a>
   </body>
