+++ test/fixtures/linked-pages/nested/team.html
@@ -1,9 +1,9 @@
 <!doctype html>
 <html>
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>Team</title>
-    <link rel="stylesheet" href="../assets/team.css">
+    <link rel="stylesheet" href="../assets/team.css" />
   </head>
   <body>
     <a href="../about.html">About</a>
